Finish doc and adapt build script.
The documentation has been completed and all intra-doc links that were
broken are now fixed. On top of that, the CI setup has been modified
to actually verify that the documentation can be built successfully.

The readme has been updated to reflect the changes and we now also have
a script called `check` that runs the checks performed by the CI on the
local machine to avoid opening pull requests that fail.