Collecting into DTS and HashTs is possible, DTS collection specialized
We have now specialized collection for `DTS` via the `collect_dts` method
and it has been optimized for ts with mapped state or edge colors and
for the case where we collect the ts itself.

Further work would require #46 to be done as this is blocking the
ability of "dropping" a color. This should be implemented in the
method [`Sproutable::collect_from`]. From there on, the changes need
to be propagated to the different `collect_...` methods on the
`Deterministic` trait.