WIP building a void type
This has a lot more consequences than anticipated. It relaxes quite a few
trait bounds but at the same time, it pulls trait bounds further to the
front. For real user facing code, this should not have too big of an
impact as it will normally work with types that do have the desirable
properties that were originally ensured by the `Color` trait that was
implemented by both edge and state color in `TransitionSystem`.

Now, if one wants to work with a generic TS, one has to manually specify
that its edge or state colors are _actual colors. With the new types
`EdgeColor` and `StateColor`, this is actually quite nice and ergonomic.